+++ buildkite/resource_cluster_test.go
@@ -14,7 +14,7 @@ func TestAccBuildkiteClusterResource(t *testing.T) {
 	basic := func(name string) string {
 		return fmt.Sprintf(`
 		provider "buildkite" {
-			timeouts {
+			timeouts = {
 				create = "10s"
 				read = "10s"
 				update = "10s"
@@ -31,7 +31,7 @@ func TestAccBuildkiteClusterResource(t *testing.T) {
 	complex := func(fields ...string) string {
 		return fmt.Sprintf(`
 		provider "buildkite" {
-			timeouts {
+			timeouts = {
 				create = "10s"
 				read = "10s"
 				update = "10s"