+++ examples/resources/buildkite_pipeline/import.sh
@@ -1,0 +1,3 @@
+# import a pipeline resource using the pipelines GraphQL ID
+# GraphQL ID for a pipeline can be found on its settings page
+terraform import buildkite_pipeline.pipeline UGlwZWxpbmUtLS00MzVjYWQ1OC1lODFkLTQ1YWYtODYzNy1iMWNmODA3MDIzOGQ=