+++ main.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/buildkite/terraform-provider-buildkite/buildkite"
 	"github.com/hashicorp/terraform-plugin-framework/providerserver"
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5/tf5server"
+	"github.com/hashicorp/terraform-plugin-go/tfprotov6/tf6server"
 )
 
 // Set at compile time from ldflags
@@ -20,15 +20,15 @@ func main() {
 	flag.BoolVar(&debug, "debug", false, "set to true to run the provider with support for debuggers like delve")
 	flag.Parse()
 
-	var serveOpts []tf5server.ServeOpt
+	var serveOpts []tf6server.ServeOpt
 
 	if debug {
-		serveOpts = append(serveOpts, tf5server.WithManagedDebug())
+		serveOpts = append(serveOpts, tf6server.WithManagedDebug())
 	}
 
-	err := tf5server.Serve(
+	err := tf6server.Serve(
 		"registry.terraform.io/buildkite/buildkite",
-		providerserver.NewProtocol5(buildkite.New(version)),
+		providerserver.NewProtocol6(buildkite.New(version)),
 		serveOpts...,
 	)